event_loop/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-appconf==1.0.3
django-bootstrap-pagination==1.7.0
django-compat==1.0.15
django-cors-headers==2.5.2
djangorestframework==3.9.2
geographiclib==1.49
geopy==1.19.0
idna==2.8
nltk==3.4
pytz==2019.1
rake-nltk==1.0.4
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
